+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/api/storage/IDrawerAttributes.java
@@ -77,4 +77,6 @@ public interface IDrawerAttributes
     default boolean isHopper () { return false; }
 
     default boolean isMagnet () { return false; }
+
+    default boolean isSuspended () { return false; }
 }