+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/BlockEntityController.java
@@ -360,6 +360,14 @@ public class BlockEntityController extends BaseBlockEntity implements IDrawerGro
         }
 
         @Override
+        public boolean setIsSuspended (boolean state) {
+            getSlotsValidAttributes(player).forEach(record -> {
+                record.setIsSuspended(state);
+            });
+            return true;
+        }
+
+        @Override
         public boolean setItemLocked (EnumSet<LockAttribute> attributes, LockAttribute attr, boolean isLocked) {
             getSlotsValidAttributes(player).forEach(record -> {
                 for (LockAttribute a : attributes)
@@ -379,6 +387,13 @@ public class BlockEntityController extends BaseBlockEntity implements IDrawerGro
         public boolean toggleIsShowingQuantity () {
             getSlotsValidAttributes(player).findFirst().ifPresent(template ->
                 setIsShowingQuantity(!template.isShowingQuantity()));
+            return true;
+        }
+
+        @Override
+        public boolean toggleIsSuspended () {
+            getSlotsValidAttributes(player).findFirst().ifPresent(template ->
+                setIsSuspended(!template.isSuspended()));
             return true;
         }
 
@@ -899,8 +914,10 @@ public class BlockEntityController extends BaseBlockEntity implements IDrawerGro
                     if (!hasAccess(candidateGroup, drawer))
                         continue;
 
-                    IDrawerAttributes attrs = getAttributes(candidateGroup);
-                    if (attrs != null && attrs.isBalancedFill())
+                    IDrawerAttributes attrs = drawer.getAttributes();
+                    if (attrs.isSuspended())
+                        continue;
+                    if (attrs.isBalancedFill())
                         rebalance.add(drawer);
 
                     if (amount == 0)
@@ -933,6 +950,10 @@ public class BlockEntityController extends BaseBlockEntity implements IDrawerGro
                         if (!hasAccess(candidateGroup, drawer))
                             continue;
 
+                        IDrawerAttributes attrs = drawer.getAttributes();
+                        if (attrs.isSuspended())
+                            continue;
+
                         amount = (simulate)
                             ? Math.max(amount - drawer.getAcceptingRemainingCapacity(), 0)
                             : drawer.adjustStoredItemCount(amount);
@@ -950,8 +971,15 @@ public class BlockEntityController extends BaseBlockEntity implements IDrawerGro
                         continue;
                     if (!testPredicateInsert(drawer, stack, predicate))
                         continue;
-                    if (!hasAccess(getGroupForDrawerSlot(slot), drawer))
+
+                    IDrawerGroup group = getGroupForDrawerSlot(slot);
+                    if (!hasAccess(group, drawer))
                         continue;
+
+                    IDrawerAttributes attrs = drawer.getAttributes();
+                    if (attrs.isSuspended())
+                        continue;
+
                     if (simulate && checkedSlots.contains(slot))
                         continue;
 
@@ -998,8 +1026,10 @@ public class BlockEntityController extends BaseBlockEntity implements IDrawerGro
                     if (!hasAccess(candidateGroup, drawer))
                         continue;
 
-                    IDrawerAttributes attrs = getAttributes(candidateGroup);
-                    if (attrs != null && attrs.isBalancedFill())
+                    IDrawerAttributes attrs = drawer.getAttributes();
+                    if (attrs.isSuspended())
+                        continue;
+                    if (attrs.isBalancedFill())
                         rebalance.add(drawer);
 
                     if (remaining == 0)
@@ -1021,6 +1051,15 @@ public class BlockEntityController extends BaseBlockEntity implements IDrawerGro
                         continue;
                     if (!testPredicateExtract(drawer, stack, predicate))
                         continue;
+
+                    IDrawerGroup group = getGroupForDrawerSlot(slot);
+                    if (!hasAccess(group, drawer))
+                        continue;
+
+                    IDrawerAttributes attrs = drawer.getAttributes();
+                    if (attrs.isSuspended())
+                        continue;
+
                     if (simulate && checkedSlots.contains(slot))
                         continue;