+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/tiledata/FractionalDrawerGroup.java
@@ -135,7 +135,7 @@ public class FractionalDrawerGroup extends BlockEntityDataShim implements IDrawe
         }
 
         @NotNull
-        IDrawerAttributes getAttributes() {
+        public IDrawerAttributes getAttributes() {
             if (cachedAttrs != null)
                 return cachedAttrs;
 
@@ -717,6 +717,11 @@ public class FractionalDrawerGroup extends BlockEntityDataShim implements IDrawe
         @Override
         public boolean isSmallestUnit() {
             return storage.isSmallestUnit(slot);
+        }
+
+        @Override
+        public @NotNull IDrawerAttributes getAttributes () {
+            return storage.getAttributes();
         }
 
         @Override